+++ BlishHud-Raid-Clears/Settings/Models/StrikeSettings.cs
@@ -8,6 +8,8 @@ public class StrikeSettings
 {
     public SettingEntry<bool> AnchorToRaidPanel { get; set; }
     public SettingEntry<StrikeComplete> StrikeCompletion { get; set; }
+    public SettingEntry<bool> StrikePanelHighlightNonWeeklyBounty { get; set; }
+    public SettingEntry<string> StrikePanelColorNonWeeklyBounty { get; set; }
     public DisplayStyle Style { get; set; }
     public GenericSettings Generic { get; set; }
 
@@ -48,7 +50,9 @@ public class StrikeSettings
 
         AnchorToRaidPanel = settings.DefineSetting(Settings.Strikes.Module.anchorToRaids);
         StrikeCompletion = settings.DefineSetting(Settings.Strikes.Module.strikeCompletion);
-        
+        StrikePanelHighlightNonWeeklyBounty = settings.DefineSetting(Settings.Strikes.Module.highlightNonWeeklyBounty);
+        StrikePanelColorNonWeeklyBounty = settings.DefineSetting(Settings.Strikes.Style.Color.nonWeeklyBounty);
+
         CleanUpOldSettings(settings);
     }