+++ BlishHud-Raid-Clears/Settings/Views/Tabs/StrikesSettingTab.cs
@@ -22,14 +22,18 @@ public class StrikesSettingTab : ISettingsMenuRegistrar
             new MenuItem(Strings.SettingsPanel_Raids_Heading_General),
             _ => new GenericGeneralView(Service.Settings.StrikeSettings.Generic, new List<SettingEntry>
             {
-                Service.Settings.StrikeSettings.StrikeCompletion,
-                Service.Settings.StrikeSettings.AnchorToRaidPanel
+                // Service.Settings.StrikeSettings.StrikeCompletion,
+                Service.Settings.StrikeSettings.AnchorToRaidPanel,
+                Service.Settings.StrikeSettings.StrikePanelHighlightNonWeeklyBounty
             })
         ));
 
         _registeredMenuItems.Add(new MenuViewItem(
             new MenuItem(Strings.SettingsPanel_Raids_Heading_Layout),
-            _ => new GenericStyleView(Service.Settings.StrikeSettings.Style, null, true)
+            _ => new GenericStyleView(Service.Settings.StrikeSettings.Style, new List<SettingEntry<string>>
+            {
+                Service.Settings.StrikeSettings.StrikePanelColorNonWeeklyBounty
+            }, true)
         ));
 
         /*_registeredMenuItems.Add(new MenuViewItem(
@@ -46,11 +50,11 @@ public class StrikesSettingTab : ISettingsMenuRegistrar
             _ => new StrikeLabelCustomizationView()
         ));
 
-        _registeredMenuItems.Add(new MenuViewItem(
-            new MenuItem(Strings.SettingsPanel_Strikes_Heading_ManageClears),
-            _ => new StrikeClearCorrectionView()
-        ));
-        
+        // Manage Clears window hidden.
+        // _registeredMenuItems.Add(new MenuViewItem(
+        //     new MenuItem(Strings.SettingsPanel_Strikes_Heading_ManageClears),
+        //     _ => new StrikeClearCorrectionView()
+        // ));
     }
     
     public IEnumerable<MenuItem> GetSettingMenus() =>